+++ apollo-router-scaffold/templates/plugin/src/plugins/{{snake_name}}.rs
@@ -1,10 +1,10 @@
 use apollo_router::plugin::Plugin;
 use apollo_router::plugin::PluginInit;
 use apollo_router::register_plugin;
-use apollo_router::stages::supergraph;
+use apollo_router::services::supergraph;
 {{#if type_basic}}
-use apollo_router::stages::execution;
-use apollo_router::stages::subgraph;
+use apollo_router::services::execution;
+use apollo_router::services::subgraph;
 {{/if}}
 {{#if type_auth}}
 use apollo_router::layers::ServiceBuilderExt;
@@ -146,7 +146,7 @@ register_plugin!("{{project_name}}", "{{snake_name}}", {{pascal_name}});
 #[cfg(test)]
 mod tests {
     use apollo_router::TestHarness;
-    use apollo_router::stages::supergraph;
+    use apollo_router::services::supergraph;
     use tower::BoxError;
     use tower::ServiceExt;
