+++ apollo-router/src/plugins/csrf.rs
@@ -14,7 +14,7 @@ use crate::layers::ServiceBuilderExt;
 use crate::plugin::Plugin;
 use crate::plugin::PluginInit;
 use crate::register_plugin;
-use crate::stages::supergraph;
+use crate::services::supergraph;
 use crate::SupergraphRequest;
 use crate::SupergraphResponse;
