+++ apollo-router/src/services/mod.rs
@@ -1,754 +1,32 @@
 //! Implementation of the various steps in the router's processing pipeline.
 
-#![allow(missing_docs)] // FIXME
-
-use std::collections::HashMap;
 use std::sync::Arc;
 
-use futures::future::ready;
-use futures::stream::once;
-use futures::stream::BoxStream;
-use futures::stream::StreamExt;
-use http::header::HeaderName;
-use http::method::Method;
-use http::HeaderValue;
-use http::StatusCode;
-use http::Uri;
-use multimap::MultiMap;
-use serde_json_bytes::ByteString;
-use static_assertions::assert_impl_all;
-use tower::BoxError;
-
 pub(crate) use self::execution_service::*;
+pub(crate) use self::query_planner::*;
 pub(crate) use self::subgraph_service::*;
 pub(crate) use self::supergraph_service::*;
-use crate::error::Error;
 use crate::graphql::Request;
-use crate::graphql::Response;
 use crate::http_ext;
-use crate::http_ext::IntoHeaderName;
-use crate::http_ext::IntoHeaderValue;
-use crate::json_ext::Object;
-use crate::json_ext::Path;
-use crate::json_ext::Value;
-use crate::query_planner::fetch::OperationKind;
-use crate::query_planner::QueryPlan;
-use crate::spec::Query;
-use crate::Context;
+pub(crate) use crate::services::execution::Request as ExecutionRequest;
+pub(crate) use crate::services::execution::Response as ExecutionResponse;
+pub(crate) use crate::services::query_planner::Request as QueryPlannerRequest;
+pub(crate) use crate::services::query_planner::Response as QueryPlannerResponse;
+pub(crate) use crate::services::subgraph::Request as SubgraphRequest;
+pub(crate) use crate::services::subgraph::Response as SubgraphResponse;
+pub(crate) use crate::services::supergraph::Request as SupergraphRequest;
+pub(crate) use crate::services::supergraph::Response as SupergraphResponse;
 
+pub mod execution;
 mod execution_service;
 pub(crate) mod layers;
 pub(crate) mod new_service;
+pub(crate) mod query_planner;
+pub mod subgraph;
 pub(crate) mod subgraph_service;
+pub mod supergraph;
 mod supergraph_service;
-
-assert_impl_all!(SupergraphRequest: Send);
-/// Represents the router processing step of the processing pipeline.
-///
-/// This consists of the parsed graphql Request, HTTP headers and contextual data for extensions.
-pub struct SupergraphRequest {
-    /// Original request to the Router.
-    pub originating_request: http_ext::Request<Request>,
-
-    /// Context for extension
-    pub context: Context,
-}
-
-impl From<http_ext::Request<Request>> for SupergraphRequest {
-    fn from(originating_request: http_ext::Request<Request>) -> Self {
-        Self {
-            originating_request,
-            context: Context::new(),
-        }
-    }
-}
-
-#[buildstructor::buildstructor]
-impl SupergraphRequest {
-    /// This is the constructor (or builder) to use when constructing a real SupergraphRequest.
-    ///
-    /// Required parameters are required in non-testing code to create a SupergraphRequest.
-    #[allow(clippy::too_many_arguments)]
-    #[builder(visibility = "pub")]
-    fn new(
-        query: Option<String>,
-        operation_name: Option<String>,
-        variables: HashMap<String, Value>,
-        extensions: HashMap<String, Value>,
-        context: Context,
-        headers: MultiMap<IntoHeaderName, IntoHeaderValue>,
-        uri: Uri,
-        method: Method,
-    ) -> Result<SupergraphRequest, BoxError> {
-        let extensions: Object = extensions
-            .into_iter()
-            .map(|(name, value)| (ByteString::from(name), value))
-            .collect();
-
-        let variables: Object = variables
-            .into_iter()
-            .map(|(name, value)| (ByteString::from(name), value))
-            .collect();
-
-        let gql_request = Request::builder()
-            .and_query(query)
-            .and_operation_name(operation_name)
-            .variables(variables)
-            .extensions(extensions)
-            .build();
-
-        let originating_request = http_ext::Request::builder()
-            .headers(headers)
-            .uri(uri)
-            .method(method)
-            .body(gql_request)
-            .build()?;
-
-        Ok(Self {
-            originating_request,
-            context,
-        })
-    }
-
-    /// This is the constructor (or builder) to use when constructing a "fake" SupergraphRequest.
-    ///
-    /// This does not enforce the provision of the data that is required for a fully functional
-    /// SupergraphRequest. It's usually enough for testing, when a fully constructed SupergraphRequest is
-    /// difficult to construct and not required for the purposes of the test.
-    ///
-    /// In addition, fake requests are expected to be valid, and will panic if given invalid values.
-    #[builder(visibility = "pub")]
-    fn fake_new(
-        query: Option<String>,
-        operation_name: Option<String>,
-        variables: HashMap<String, Value>,
-        extensions: HashMap<String, Value>,
-        context: Option<Context>,
-        mut headers: MultiMap<IntoHeaderName, IntoHeaderValue>,
-    ) -> Result<SupergraphRequest, BoxError> {
-        // Avoid testing requests getting blocked by the CSRF-prevention plugin
-        headers
-            .entry(IntoHeaderName::HeaderName(http::header::CONTENT_TYPE))
-            .or_insert(IntoHeaderValue::HeaderValue(HeaderValue::from_static(
-                "application/json",
-            )));
-        SupergraphRequest::new(
-            query,
-            operation_name,
-            variables,
-            extensions,
-            context.unwrap_or_default(),
-            headers,
-            Uri::from_static("http://default"),
-            Method::GET,
-        )
-    }
-
-    /// Create a request with an example query, for tests
-    #[builder(visibility = "pub")]
-    fn canned_new(
-        operation_name: Option<String>,
-        extensions: HashMap<String, Value>,
-        context: Option<Context>,
-        headers: MultiMap<IntoHeaderName, IntoHeaderValue>,
-    ) -> Result<SupergraphRequest, BoxError> {
-        let query = "
-            query TopProducts($first: Int) { 
-                topProducts(first: $first) { 
-                    upc 
-                    name 
-                    reviews { 
-                        id 
-                        product { name } 
-                        author { id name } 
-                    } 
-                } 
-            }
-        ";
-        let mut variables = HashMap::new();
-        variables.insert("first".to_owned(), 2_usize.into());
-        Self::fake_new(
-            Some(query.to_owned()),
-            operation_name,
-            variables,
-            extensions,
-            context,
-            headers,
-        )
-    }
-}
-
-assert_impl_all!(SupergraphResponse: Send);
-/// [`Context`] and [`http_ext::Response<Response>`] for the response.
-///
-/// This consists of the response body and the context.
-pub struct SupergraphResponse {
-    pub response: http_ext::Response<BoxStream<'static, Response>>,
-    pub context: Context,
-}
-
-#[buildstructor::buildstructor]
-impl SupergraphResponse {
-    /// This is the constructor (or builder) to use when constructing a real SupergraphResponse..
-    ///
-    /// Required parameters are required in non-testing code to create a SupergraphResponse..
-    #[allow(clippy::too_many_arguments)]
-    #[builder(visibility = "pub")]
-    fn new(
-        data: Option<Value>,
-        path: Option<Path>,
-        errors: Vec<Error>,
-        extensions: HashMap<String, Value>,
-        status_code: Option<StatusCode>,
-        headers: MultiMap<IntoHeaderName, IntoHeaderValue>,
-        context: Context,
-    ) -> Result<Self, BoxError> {
-        let extensions: Object = extensions
-            .into_iter()
-            .map(|(name, value)| (ByteString::from(name), value))
-            .collect();
-        // Build a response
-        let b = Response::builder()
-            .and_path(path)
-            .errors(errors)
-            .extensions(extensions);
-        let res = match data {
-            Some(data) => b.data(data).build(),
-            None => b.build(),
-        };
-
-        // Build an http Response
-        let mut builder = http::Response::builder().status(status_code.unwrap_or(StatusCode::OK));
-        for (key, values) in headers {
-            let header_name: HeaderName = key.try_into()?;
-            for value in values {
-                let header_value: HeaderValue = value.try_into()?;
-                builder = builder.header(header_name.clone(), header_value);
-            }
-        }
-
-        let http_response = builder.body(once(ready(res)).boxed())?;
-
-        // Create a compatible Response
-        let compat_response = http_ext::Response {
-            inner: http_response,
-        };
-
-        Ok(Self {
-            response: compat_response,
-            context,
-        })
-    }
-
-    /// This is the constructor (or builder) to use when constructing a "fake" SupergraphResponse.
-    ///
-    /// This does not enforce the provision of the data that is required for a fully functional
-    /// SupergraphResponse. It's usually enough for testing, when a fully constructed SupergraphResponse is
-    /// difficult to construct and not required for the purposes of the test.
-    ///
-    /// In addition, fake responses are expected to be valid, and will panic if given invalid values.
-    #[allow(clippy::too_many_arguments)]
-    #[builder(visibility = "pub")]
-    fn fake_new(
-        data: Option<Value>,
-        path: Option<Path>,
-        errors: Vec<Error>,
-        extensions: HashMap<String, Value>,
-        status_code: Option<StatusCode>,
-        headers: MultiMap<IntoHeaderName, IntoHeaderValue>,
-        context: Option<Context>,
-    ) -> Result<Self, BoxError> {
-        SupergraphResponse::new(
-            data,
-            path,
-            errors,
-            extensions,
-            status_code,
-            headers,
-            context.unwrap_or_default(),
-        )
-    }
-
-    /// This is the constructor (or builder) to use when constructing a SupergraphResponse that represents a global error.
-    /// It has no path and no response data.
-    /// This is useful for things such as authentication errors.
-    #[builder(visibility = "pub")]
-    fn error_new(
-        errors: Vec<Error>,
-        status_code: Option<StatusCode>,
-        headers: MultiMap<IntoHeaderName, IntoHeaderValue>,
-        context: Context,
-    ) -> Result<Self, BoxError> {
-        SupergraphResponse::new(
-            Default::default(),
-            None,
-            errors,
-            Default::default(),
-            status_code,
-            headers,
-            context,
-        )
-    }
-
-    pub fn new_from_graphql_response(response: Response, context: Context) -> Self {
-        Self {
-            response: http::Response::new(once(ready(response)).boxed()).into(),
-            context,
-        }
-    }
-}
-
-impl SupergraphResponse {
-    pub async fn next_response(&mut self) -> Option<Response> {
-        self.response.body_mut().next().await
-    }
-
-    pub fn new_from_response(
-        response: http_ext::Response<BoxStream<'static, Response>>,
-        context: Context,
-    ) -> Self {
-        Self { response, context }
-    }
-
-    pub fn map<F>(self, f: F) -> SupergraphResponse
-    where
-        F: FnOnce(BoxStream<'static, Response>) -> BoxStream<'static, Response>,
-    {
-        SupergraphResponse {
-            context: self.context,
-            response: self.response.map(f),
-        }
-    }
-
-    pub fn map_stream(self, f: impl FnMut(Response) -> Response + Send + 'static) -> Self {
-        self.map(move |stream| stream.map(f).boxed())
-    }
-}
-
-assert_impl_all!(QueryPlannerRequest: Send);
-/// [`Context`] for the request.
-#[derive(Clone, Debug)]
-pub(crate) struct QueryPlannerRequest {
-    pub(crate) query: String,
-    pub(crate) operation_name: Option<String>,
-    pub(crate) context: Context,
-}
-
-#[buildstructor::buildstructor]
-impl QueryPlannerRequest {
-    /// This is the constructor (or builder) to use when constructing a real QueryPlannerRequest.
-    ///
-    /// Required parameters are required in non-testing code to create a QueryPlannerRequest.
-    #[builder]
-    pub(crate) fn new(
-        query: String,
-        operation_name: Option<String>,
-        context: Context,
-    ) -> QueryPlannerRequest {
-        Self {
-            query,
-            operation_name,
-            context,
-        }
-    }
-}
-
-assert_impl_all!(QueryPlannerResponse: Send);
-/// [`Context`] and [`QueryPlan`] for the response.
-pub(crate) struct QueryPlannerResponse {
-    pub(crate) content: QueryPlannerContent,
-    pub(crate) context: Context,
-}
-
-/// Query, QueryPlan and Introspection data.
-#[derive(Debug, Clone)]
-pub enum QueryPlannerContent {
-    Plan {
-        query: Arc<Query>,
-        plan: Arc<QueryPlan>,
-    },
-    Introspection {
-        response: Box<Response>,
-    },
-    IntrospectionDisabled,
-}
-
-#[buildstructor::buildstructor]
-impl QueryPlannerResponse {
-    /// This is the constructor (or builder) to use when constructing a real QueryPlannerResponse.
-    ///
-    /// Required parameters are required in non-testing code to create a QueryPlannerResponse.
-    #[builder]
-    pub(crate) fn new(content: QueryPlannerContent, context: Context) -> QueryPlannerResponse {
-        Self { content, context }
-    }
-}
-
-assert_impl_all!(SubgraphRequest: Send);
-/// [`Context`], [`OperationKind`] and [`http_ext::Request<Request>`] for the request.
-pub struct SubgraphRequest {
-    /// Original request to the Router.
-    pub originating_request: Arc<http_ext::Request<Request>>,
-
-    pub subgraph_request: http_ext::Request<Request>,
-
-    pub operation_kind: OperationKind,
-
-    pub context: Context,
-}
-
-#[buildstructor::buildstructor]
-impl SubgraphRequest {
-    /// This is the constructor (or builder) to use when constructing a real SubgraphRequest.
-    ///
-    /// Required parameters are required in non-testing code to create a SubgraphRequest.
-    #[builder(visibility = "pub")]
-    fn new(
-        originating_request: Arc<http_ext::Request<Request>>,
-        subgraph_request: http_ext::Request<Request>,
-        operation_kind: OperationKind,
-        context: Context,
-    ) -> SubgraphRequest {
-        Self {
-            originating_request,
-            subgraph_request,
-            operation_kind,
-            context,
-        }
-    }
-
-    /// This is the constructor (or builder) to use when constructing a "fake" SubgraphRequest.
-    ///
-    /// This does not enforce the provision of the data that is required for a fully functional
-    /// SubgraphRequest. It's usually enough for testing, when a fully consructed SubgraphRequest is
-    /// difficult to construct and not required for the pusposes of the test.
-    #[builder(visibility = "pub")]
-    fn fake_new(
-        originating_request: Option<Arc<http_ext::Request<Request>>>,
-        subgraph_request: Option<http_ext::Request<Request>>,
-        operation_kind: Option<OperationKind>,
-        context: Option<Context>,
-    ) -> SubgraphRequest {
-        SubgraphRequest::new(
-            originating_request.unwrap_or_else(|| {
-                Arc::new(
-                    http_ext::Request::fake_builder()
-                        .headers(Default::default())
-                        .body(Default::default())
-                        .build()
-                        .expect("fake builds should always work; qed"),
-                )
-            }),
-            subgraph_request.unwrap_or_else(|| {
-                http_ext::Request::fake_builder()
-                    .headers(Default::default())
-                    .body(Default::default())
-                    .build()
-                    .expect("fake builds should always work; qed")
-            }),
-            operation_kind.unwrap_or(OperationKind::Query),
-            context.unwrap_or_default(),
-        )
-    }
-}
-
-assert_impl_all!(SubgraphResponse: Send);
-/// [`Context`] and [`http_ext::Response<Response>`] for the response.
-///
-/// This consists of the subgraph response and the context.
-#[derive(Clone, Debug)]
-pub struct SubgraphResponse {
-    pub response: http_ext::Response<Response>,
-
-    pub context: Context,
-}
-
-#[buildstructor::buildstructor]
-impl SubgraphResponse {
-    /// This is the constructor to use when constructing a real SubgraphResponse..
-    ///
-    /// In this case, you already have a valid response and just wish to associate it with a context
-    /// and create a SubgraphResponse.
-    pub fn new_from_response(
-        response: http_ext::Response<Response>,
-        context: Context,
-    ) -> SubgraphResponse {
-        Self { response, context }
-    }
-
-    /// This is the constructor (or builder) to use when constructing a real SubgraphResponse.
-    ///
-    /// The parameters are not optional, because in a live situation all of these properties must be
-    /// set and be correct to create a SubgraphResponse.
-    #[builder(visibility = "pub")]
-    fn new(
-        label: Option<String>,
-        data: Option<Value>,
-        path: Option<Path>,
-        errors: Vec<Error>,
-        extensions: Object,
-        status_code: Option<StatusCode>,
-        context: Context,
-    ) -> SubgraphResponse {
-        // Build a response
-        let res = Response::builder()
-            .and_label(label)
-            .data(data.unwrap_or_default())
-            .and_path(path)
-            .errors(errors)
-            .extensions(extensions)
-            .build();
-
-        // Build an http Response
-        let http_response = http::Response::builder()
-            .status(status_code.unwrap_or(StatusCode::OK))
-            .body(res)
-            .expect("Response is serializable; qed");
-
-        // Create a compatible Response
-        let compat_response = http_ext::Response {
-            inner: http_response,
-        };
-
-        Self {
-            response: compat_response,
-            context,
-        }
-    }
-
-    /// This is the constructor (or builder) to use when constructing a "fake" SubgraphResponse.
-    ///
-    /// This does not enforce the provision of the data that is required for a fully functional
-    /// SubgraphResponse. It's usually enough for testing, when a fully consructed SubgraphResponse is
-    /// difficult to construct and not required for the pusposes of the test.
-    #[builder(visibility = "pub")]
-    fn fake_new(
-        label: Option<String>,
-        data: Option<Value>,
-        path: Option<Path>,
-        errors: Vec<Error>,
-        extensions: Option<Object>,
-        status_code: Option<StatusCode>,
-        context: Option<Context>,
-    ) -> SubgraphResponse {
-        SubgraphResponse::new(
-            label,
-            data,
-            path,
-            errors,
-            extensions.unwrap_or_default(),
-            status_code,
-            context.unwrap_or_default(),
-        )
-    }
-
-    /// This is the constructor (or builder) to use when constructing a SubgraphResponse that represents a global error.
-    /// It has no path and no response data.
-    /// This is useful for things such as authentication errors.
-    #[builder(visibility = "pub")]
-    fn error_new(
-        errors: Vec<Error>,
-        status_code: Option<StatusCode>,
-        context: Context,
-    ) -> Result<SubgraphResponse, BoxError> {
-        Ok(SubgraphResponse::new(
-            Default::default(),
-            Default::default(),
-            Default::default(),
-            errors,
-            Default::default(),
-            status_code,
-            context,
-        ))
-    }
-}
-
-assert_impl_all!(ExecutionRequest: Send);
-/// [`Context`] and [`QueryPlan`] for the request.
-pub struct ExecutionRequest {
-    /// Original request to the Router.
-    pub originating_request: http_ext::Request<Request>,
-
-    pub query_plan: Arc<QueryPlan>,
-
-    pub context: Context,
-}
-
-#[buildstructor::buildstructor]
-impl ExecutionRequest {
-    /// This is the constructor (or builder) to use when constructing a real ExecutionRequest.
-    ///
-    /// The parameters are not optional, because in a live situation all of these properties must be
-    /// set and be correct to create a ExecutionRequest.
-    #[builder(visibility = "pub")]
-    fn new(
-        originating_request: http_ext::Request<Request>,
-        query_plan: Arc<QueryPlan>,
-        context: Context,
-    ) -> ExecutionRequest {
-        Self {
-            originating_request,
-            query_plan,
-            context,
-        }
-    }
-
-    /// This is the constructor (or builder) to use when constructing a "fake" ExecutionRequest.
-    ///
-    /// This does not enforce the provision of the data that is required for a fully functional
-    /// ExecutionRequest. It's usually enough for testing, when a fully consructed ExecutionRequest is
-    /// difficult to construct and not required for the pusposes of the test.
-    #[builder(visibility = "pub")]
-    fn fake_new(
-        originating_request: Option<http_ext::Request<Request>>,
-        query_plan: Option<QueryPlan>,
-        context: Option<Context>,
-    ) -> ExecutionRequest {
-        ExecutionRequest::new(
-            originating_request.unwrap_or_else(|| {
-                http_ext::Request::fake_builder()
-                    .headers(Default::default())
-                    .body(Default::default())
-                    .build()
-                    .expect("fake builds should always work; qed")
-            }),
-            Arc::new(query_plan.unwrap_or_else(|| QueryPlan::fake_builder().build())),
-            context.unwrap_or_default(),
-        )
-    }
-}
-
-assert_impl_all!(ExecutionResponse: Send);
-/// [`Context`] and [`http_ext::Response<Response>`] for the response.
-///
-/// This consists of the execution response and the context.
-pub struct ExecutionResponse {
-    pub response: http_ext::Response<BoxStream<'static, Response>>,
-
-    pub context: Context,
-}
-
-#[buildstructor::buildstructor]
-impl ExecutionResponse {
-    /// This is the constructor (or builder) to use when constructing a real SupergraphRequest.
-    ///
-    /// The parameters are not optional, because in a live situation all of these properties must be
-    /// set and be correct to create a SupergraphRequest.
-    #[builder(visibility = "pub")]
-    fn new(
-        label: Option<String>,
-        data: Option<Value>,
-        path: Option<Path>,
-        errors: Vec<Error>,
-        extensions: Object,
-        status_code: Option<StatusCode>,
-        context: Context,
-    ) -> Self {
-        // Build a response
-        let res = Response::builder()
-            .and_label(label)
-            .data(data.unwrap_or_default())
-            .and_path(path)
-            .errors(errors)
-            .extensions(extensions)
-            .build();
-
-        // Build an http Response
-        let http_response = http::Response::builder()
-            .status(status_code.unwrap_or(StatusCode::OK))
-            .body(once(ready(res)).boxed())
-            .expect("Response is serializable; qed");
-
-        // Create a compatible Response
-        let compat_response = http_ext::Response {
-            inner: http_response,
-        };
-
-        Self {
-            response: compat_response,
-            context,
-        }
-    }
-
-    /// This is the constructor (or builder) to use when constructing a "fake" ExecutionResponse.
-    ///
-    /// This does not enforce the provision of the data that is required for a fully functional
-    /// ExecutionResponse. It's usually enough for testing, when a fully consructed
-    /// ExecutionResponse is difficult to construct and not required for the pusposes of the test.
-    #[builder(visibility = "pub")]
-    fn fake_new(
-        label: Option<String>,
-        data: Option<Value>,
-        path: Option<Path>,
-        errors: Vec<Error>,
-        extensions: Option<Object>,
-        status_code: Option<StatusCode>,
-        context: Option<Context>,
-    ) -> Self {
-        ExecutionResponse::new(
-            label,
-            data,
-            path,
-            errors,
-            extensions.unwrap_or_default(),
-            status_code,
-            context.unwrap_or_default(),
-        )
-    }
-
-    /// This is the constructor (or builder) to use when constructing a ExecutionResponse that represents a global error.
-    /// It has no path and no response data.
-    /// This is useful for things such as authentication errors.
-    #[allow(unused_variables)]
-    #[builder(visibility = "pub")]
-    fn error_new(
-        errors: Vec<Error>,
-        status_code: Option<StatusCode>,
-        headers: MultiMap<IntoHeaderName, IntoHeaderValue>,
-        context: Context,
-    ) -> Result<Self, BoxError> {
-        Ok(ExecutionResponse::new(
-            Default::default(),
-            Default::default(),
-            Default::default(),
-            errors,
-            Default::default(),
-            status_code,
-            context,
-        ))
-    }
-}
-
-impl ExecutionResponse {
-    /// This is the constructor to use when constructing a real ExecutionResponse.
-    ///
-    /// In this case, you already have a valid request and just wish to associate it with a context
-    /// and create a ExecutionResponse.
-    pub fn new_from_response(
-        response: http_ext::Response<BoxStream<'static, Response>>,
-        context: Context,
-    ) -> Self {
-        Self { response, context }
-    }
-
-    pub fn map<F>(self, f: F) -> ExecutionResponse
-    where
-        F: FnOnce(BoxStream<'static, Response>) -> BoxStream<'static, Response>,
-    {
-        ExecutionResponse {
-            context: self.context,
-            response: self.response.map(f),
-        }
-    }
-
-    pub fn map_stream(self, f: impl FnMut(Response) -> Response + Send + 'static) -> Self {
-        self.map(move |stream| stream.map(f).boxed())
-    }
-
-    pub async fn next_response(&mut self) -> Option<Response> {
-        self.response.body_mut().next().await
-    }
-}
+pub mod transport;
 
 impl AsRef<Request> for http_ext::Request<Request> {
     fn as_ref(&self) -> &Request {
@@ -759,110 +37,5 @@ impl AsRef<Request> for http_ext::Request<Request> {
 impl AsRef<Request> for Arc<http_ext::Request<Request>> {
     fn as_ref(&self) -> &Request {
         self.body()
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use http::HeaderValue;
-    use http::Method;
-    use http::Uri;
-    use serde_json::json;
-
-    use crate::graphql;
-    use crate::Context;
-    use crate::SupergraphRequest;
-    use crate::SupergraphResponse;
-
-    #[test]
-    fn supergraph_request_builder() {
-        let request = SupergraphRequest::builder()
-            .header("a", "b")
-            .header("a", "c")
-            .uri(Uri::from_static("http://example.com"))
-            .method(Method::POST)
-            .query("query { topProducts }")
-            .operation_name("Default")
-            .context(Context::new())
-            // We need to follow up on this. How can users creat this easily?
-            .extension("foo", json!({}))
-            // We need to follow up on this. How can users creat this easily?
-            .variable("bar", json!({}))
-            .build()
-            .unwrap();
-        assert_eq!(
-            request
-                .originating_request
-                .headers()
-                .get_all("a")
-                .into_iter()
-                .collect::<Vec<_>>(),
-            vec![HeaderValue::from_static("b"), HeaderValue::from_static("c")]
-        );
-        assert_eq!(
-            request.originating_request.uri(),
-            &Uri::from_static("http://example.com")
-        );
-        assert_eq!(
-            request.originating_request.body().extensions.get("foo"),
-            Some(&json!({}).into())
-        );
-        assert_eq!(
-            request.originating_request.body().variables.get("bar"),
-            Some(&json!({}).into())
-        );
-        assert_eq!(request.originating_request.method(), Method::POST);
-
-        let extensions = serde_json_bytes::Value::from(json!({"foo":{}}))
-            .as_object()
-            .unwrap()
-            .clone();
-
-        let variables = serde_json_bytes::Value::from(json!({"bar":{}}))
-            .as_object()
-            .unwrap()
-            .clone();
-        assert_eq!(
-            request.originating_request.body(),
-            &graphql::Request::builder()
-                .variables(variables)
-                .extensions(extensions)
-                .operation_name("Default")
-                .query("query { topProducts }")
-                .build()
-        );
-    }
-
-    #[tokio::test]
-    async fn supergraph_response_builder() {
-        let mut response = SupergraphResponse::builder()
-            .header("a", "b")
-            .header("a", "c")
-            .context(Context::new())
-            .extension("foo", json!({}))
-            .data(json!({}))
-            .build()
-            .unwrap();
-
-        assert_eq!(
-            response
-                .response
-                .headers()
-                .get_all("a")
-                .into_iter()
-                .collect::<Vec<_>>(),
-            vec![HeaderValue::from_static("b"), HeaderValue::from_static("c")]
-        );
-        let extensions = serde_json_bytes::Value::from(json!({"foo":{}}))
-            .as_object()
-            .unwrap()
-            .clone();
-        assert_eq!(
-            response.next_response().await.unwrap(),
-            graphql::Response::builder()
-                .extensions(extensions)
-                .data(json!({}))
-                .build()
-        );
     }
 }
