+++ apollo-router/src/services/subgraph.rs
@@ -1,0 +1,202 @@
+#![allow(missing_docs)] // FIXME
+
+use std::sync::Arc;
+
+use http::StatusCode;
+use serde_json_bytes::ByteString;
+use serde_json_bytes::Map as JsonMap;
+use serde_json_bytes::Value;
+use static_assertions::assert_impl_all;
+use tower::BoxError;
+
+use crate::error::Error;
+use crate::graphql;
+use crate::http_ext;
+use crate::json_ext::Object;
+use crate::json_ext::Path;
+use crate::query_planner::fetch::OperationKind;
+use crate::Context;
+
+pub type BoxService = tower::util::BoxService<Request, Response, BoxError>;
+pub type BoxCloneService = tower::util::BoxCloneService<Request, Response, BoxError>;
+pub type ServiceResult = Result<Response, BoxError>;
+
+assert_impl_all!(Request: Send);
+/// [`Context`], [`OperationKind`] and [`http_ext::Request<Request>`] for the request.
+pub struct Request {
+    /// Original request to the Router.
+    pub originating_request: Arc<http_ext::Request<graphql::Request>>,
+
+    pub subgraph_request: http_ext::Request<graphql::Request>,
+
+    pub operation_kind: OperationKind,
+
+    pub context: Context,
+}
+
+#[buildstructor::buildstructor]
+impl Request {
+    /// This is the constructor (or builder) to use when constructing a real Request.
+    ///
+    /// Required parameters are required in non-testing code to create a Request.
+    #[builder(visibility = "pub")]
+    fn new(
+        originating_request: Arc<http_ext::Request<graphql::Request>>,
+        subgraph_request: http_ext::Request<graphql::Request>,
+        operation_kind: OperationKind,
+        context: Context,
+    ) -> Request {
+        Self {
+            originating_request,
+            subgraph_request,
+            operation_kind,
+            context,
+        }
+    }
+
+    /// This is the constructor (or builder) to use when constructing a "fake" Request.
+    ///
+    /// This does not enforce the provision of the data that is required for a fully functional
+    /// Request. It's usually enough for testing, when a fully consructed Request is
+    /// difficult to construct and not required for the pusposes of the test.
+    #[builder(visibility = "pub")]
+    fn fake_new(
+        originating_request: Option<Arc<http_ext::Request<graphql::Request>>>,
+        subgraph_request: Option<http_ext::Request<graphql::Request>>,
+        operation_kind: Option<OperationKind>,
+        context: Option<Context>,
+    ) -> Request {
+        Request::new(
+            originating_request.unwrap_or_else(|| {
+                Arc::new(
+                    http_ext::Request::fake_builder()
+                        .headers(Default::default())
+                        .body(Default::default())
+                        .build()
+                        .expect("fake builds should always work; qed"),
+                )
+            }),
+            subgraph_request.unwrap_or_else(|| {
+                http_ext::Request::fake_builder()
+                    .headers(Default::default())
+                    .body(Default::default())
+                    .build()
+                    .expect("fake builds should always work; qed")
+            }),
+            operation_kind.unwrap_or(OperationKind::Query),
+            context.unwrap_or_default(),
+        )
+    }
+}
+
+assert_impl_all!(Response: Send);
+/// [`Context`] and [`http_ext::Response<graphql::Response>`] for the response.
+///
+/// This consists of the subgraph response and the context.
+#[derive(Clone, Debug)]
+pub struct Response {
+    pub response: http_ext::Response<graphql::Response>,
+
+    pub context: Context,
+}
+
+#[buildstructor::buildstructor]
+impl Response {
+    /// This is the constructor to use when constructing a real Response..
+    ///
+    /// In this case, you already have a valid response and just wish to associate it with a context
+    /// and create a Response.
+    pub fn new_from_response(
+        response: http_ext::Response<graphql::Response>,
+        context: Context,
+    ) -> Response {
+        Self { response, context }
+    }
+
+    /// This is the constructor (or builder) to use when constructing a real Response.
+    ///
+    /// The parameters are not optional, because in a live situation all of these properties must be
+    /// set and be correct to create a Response.
+    #[builder(visibility = "pub")]
+    fn new(
+        label: Option<String>,
+        data: Option<Value>,
+        path: Option<Path>,
+        errors: Vec<Error>,
+        extensions: Object,
+        status_code: Option<StatusCode>,
+        context: Context,
+    ) -> Response {
+        // Build a response
+        let res = graphql::Response::builder()
+            .and_label(label)
+            .data(data.unwrap_or_default())
+            .and_path(path)
+            .errors(errors)
+            .extensions(extensions)
+            .build();
+
+        // Build an http Response
+        let http_response = http::Response::builder()
+            .status(status_code.unwrap_or(StatusCode::OK))
+            .body(res)
+            .expect("Response is serializable; qed");
+
+        // Create a compatible Response
+        let compat_response = http_ext::Response {
+            inner: http_response,
+        };
+
+        Self {
+            response: compat_response,
+            context,
+        }
+    }
+
+    /// This is the constructor (or builder) to use when constructing a "fake" Response.
+    ///
+    /// This does not enforce the provision of the data that is required for a fully functional
+    /// Response. It's usually enough for testing, when a fully consructed Response is
+    /// difficult to construct and not required for the pusposes of the test.
+    #[builder(visibility = "pub")]
+    fn fake_new(
+        label: Option<String>,
+        data: Option<Value>,
+        path: Option<Path>,
+        errors: Vec<Error>,
+        // Skip the `Object` type alias in order to use buildstructor’s map special-casing
+        extensions: JsonMap<ByteString, Value>,
+        status_code: Option<StatusCode>,
+        context: Option<Context>,
+    ) -> Response {
+        Response::new(
+            label,
+            data,
+            path,
+            errors,
+            extensions,
+            status_code,
+            context.unwrap_or_default(),
+        )
+    }
+
+    /// This is the constructor (or builder) to use when constructing a Response that represents a global error.
+    /// It has no path and no response data.
+    /// This is useful for things such as authentication errors.
+    #[builder(visibility = "pub")]
+    fn error_new(
+        errors: Vec<Error>,
+        status_code: Option<StatusCode>,
+        context: Context,
+    ) -> Result<Response, BoxError> {
+        Ok(Response::new(
+            Default::default(),
+            Default::default(),
+            Default::default(),
+            errors,
+            Default::default(),
+            status_code,
+            context,
+        ))
+    }
+}
