+++ apollo-router/tests/integration_tests.rs
@@ -12,8 +12,8 @@ use apollo_router::graphql::Request;
 use apollo_router::http_ext;
 use apollo_router::plugin::Plugin;
 use apollo_router::plugin::PluginInit;
-use apollo_router::stages::subgraph;
-use apollo_router::stages::supergraph;
+use apollo_router::services::subgraph;
+use apollo_router::services::supergraph;
 use apollo_router::Context;
 use apollo_router::_private::TelemetryPlugin;
 use http::Method;
