+++ examples/forbid-anonymous-operations/src/forbid_anonymous_operations.rs
@@ -5,7 +5,7 @@ use apollo_router::layers::ServiceBuilderExt;
 use apollo_router::plugin::Plugin;
 use apollo_router::plugin::PluginInit;
 use apollo_router::register_plugin;
-use apollo_router::stages::supergraph;
+use apollo_router::services::supergraph;
 use http::StatusCode;
 use tower::BoxError;
 use tower::ServiceBuilder;
@@ -94,7 +94,7 @@ mod tests {
     use apollo_router::graphql;
     use apollo_router::plugin::test;
     use apollo_router::plugin::Plugin;
-    use apollo_router::stages::supergraph;
+    use apollo_router::services::supergraph;
     use http::StatusCode;
     use serde_json::json;
     use tower::ServiceExt;
