+++ examples/op-name-to-header/src/main.rs
@@ -13,7 +13,7 @@ fn main() -> Result<()> {
 #[cfg(test)]
 mod tests {
     use apollo_router::plugin::test;
-    use apollo_router::stages::supergraph;
+    use apollo_router::services::supergraph;
     use http::StatusCode;
     use tower::util::ServiceExt;
