+++ examples/rhai-logging/src/main.rs
@@ -14,7 +14,7 @@ fn main() -> Result<()> {
 #[cfg(test)]
 mod tests {
     use apollo_router::plugin::test;
-    use apollo_router::stages::supergraph;
+    use apollo_router::services::supergraph;
     use http::StatusCode;
     use tower::util::ServiceExt;
