+++ examples/supergraph_sdl/src/supergraph_sdl.rs
@@ -4,7 +4,7 @@ use apollo_compiler::ApolloCompiler;
 use apollo_router::plugin::Plugin;
 use apollo_router::plugin::PluginInit;
 use apollo_router::register_plugin;
-use apollo_router::stages::supergraph;
+use apollo_router::services::supergraph;
 use tower::BoxError;
 use tower::ServiceBuilder;
 use tower::ServiceExt;
